requirements.txt
kafka-python-ng
pyspark
cassandra-driver
six
faker